+++ src/jaws_site/transfers.py
@@ -2,6 +2,7 @@
 Transfer is a collection of files/folders to transfer (e.g. via Globus, FTP, etc.)
 Items are stored in a relational database.
 """
+
 import concurrent.futures
 import json
 import logging
@@ -23,16 +24,20 @@ MAX_ERROR_STRING_LEN = 1024
 
 
 def mkdir(path, mode=None):
-    if mode is None:
-        mode = int(config.conf.get("SITE", "folder_permissions", "775"), base=8)
-    if os.path.isdir(path):
-        os.chmod(path, mode)
-    else:
-        (head, tail) = os.path.split(path)
-        mkdir(head, mode)
-        if not os.path.exists(path):
-            os.mkdir(path)
-            os.chmod(path, mode)
+    mode = int(config.conf.get("SITE", "folder_permissions", "777"), base=8)
+
+    # Split the path into components, skip the first element if it's empty
+    path_parts = path.strip("/").split("/")
+
+    # Reconstruct the path and create directories
+    current_path = "/"
+    for part in path_parts:
+        current_path = os.path.join(current_path, part)
+
+        # Create the directory if it doesn't exist and set permissions
+        if not os.path.exists(current_path):
+            os.makedirs(current_path, mode=mode, exist_ok=True)
+            os.chmod(current_path, mode)
 
 
 class TransferError(Exception):
@@ -531,10 +536,7 @@ class FixPerms:
     def from_params(cls, session, params):
         """Create new transfer from parameter values and save in RDb."""
         try:
-            data = models.Fix_Perms(
-                base_dir=params["base_dir"],
-                status="queued"
-            )
+            data = models.Fix_Perms(base_dir=params["base_dir"], status="queued")
         except SQLAlchemyError as error:
             raise FixPermsValueError(
                 f"Error creating model for new FixPerms: {params}: {error}"